Move component ConfirmedCasesNumberCard from 'components/cards' to 'components/index/CardsMonitoring'
